requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
albucore==0.0.24
albumentations==0.4.6
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.0
blinker==1.9.0
braceexpand==0.1.7
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
defusedxml==0.7.1
diffusers==0.17.0
dotenv==0.9.9
easydict==1.13
exceptiongroup==1.3.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio==3.41.2
gradio_client==0.5.0
groovy==0.1.2
-e git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44#egg=groundingdino
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.13.3
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kornia==0.6.10
lazy_loader==0.4
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
narwhals==1.46.0
networkx==3.3
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
orjson==3.10.18
packaging==25.0
Padding==0.5
pandas==2.3.0
pillow==10.4.0
platformdirs==4.3.8
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.7.0
scipy==1.15.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.4.10
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
supervision==0.25.1
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.16
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchmetrics==1.7.4
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.28.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webdataset==1.0.2
websockets==11.0.3
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
